fix: Add role to token

+++ src/main/java/com/mirela/medicalrecordapp/service/impl/UserServiceImpl.java
@@ -52,7 +52,7 @@ public class UserServiceImpl implements UserService {
                 .email(userRequest.getEmail())
                 .password(passwordEncoder.encode(userRequest.getPassword()))
                 .phoneNumber(userRequest.getPhoneNumber())
-                .role(Role.ADMIN)
+                .role(userRequest.getRole())
                 .build();
 
         return userRepository.save(user);
@@ -71,6 +71,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long id) {
-    userRepository.deleteById(id);
+        User user = userRepository.findById(id).orElse(null);
+        if (user != null) {
+            userRepository.deleteById(id);
+        }
     }
 }
